# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ChemAI Model Extractor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /share/model-cards/{extractionId}:
    get:
      operationId: getPublicModelCard
      tags: [share]
      summary: Get a publicly shared model card (no auth required)
      description: >-
        Returns the model card for public sharing. The project must have
        visibility=public. Audit/provider fields (systemPrompt,
        rawProviderResponse, promptTemplateSummary) are redacted in the response.
      parameters:
        - name: extractionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Public model card with audit fields redacted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelCard"
        "404":
          description: Model card not found or not publicly shared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Equation:
      type: object
      properties:
        id:
          type: integer
        ordinal:
          type: integer
        label:
          type: string
          description: Short label, e.g. "(1)" or "Eq. 3".
        latex:
          type: string
          description: LaTeX representation of the equation.
        plaintext:
          type: string
          description: Plain-text (ASCII) representation.
        meaning:
          type: string
          description: Human-readable explanation of what the equation represents.
        variablesInvolved:
          type: array
          items:
            type: string
          description: Symbols appearing in this equation.
        confidence:
          type: string
          enum: [high, medium, low]
        description:
          type: string
          description: Auto-generated combined description (label + meaning + plaintext). Kept for backward compatibility.
        sourceQuote:
          type: string
        editedByUser:
          type: boolean
          description: True if any field has been manually edited after extraction.
        originalValue:
          type: ["object", "null"]
          additionalProperties: true
          description: Snapshot of the row before the first human edit. Null if unedited.
      required: [id, ordinal, label, latex, plaintext, meaning, variablesInvolved, confidence, description, sourceQuote, editedByUser]
    Variable:
      type: object
      properties:
        id:
          type: integer
        ordinal:
          type: integer
        symbol:
          type: string
        name:
          type: string
        meaning:
          type: string
          description: Explanation of what the variable represents physically.
        unit:
          type: string
        role:
          type: string
          enum: [state, input, output]
        confidence:
          type: string
          enum: [high, medium, low]
        sourceQuote:
          type: string
        editedByUser:
          type: boolean
        originalValue:
          type: ["object", "null"]
          additionalProperties: true
      required: [id, ordinal, symbol, name, meaning, unit, role, confidence, sourceQuote, editedByUser]
    Parameter:
      type: object
      properties:
        id:
          type: integer
        ordinal:
          type: integer
        symbol:
          type: string
        name:
          type: string
          description: Full descriptive name of the parameter.
        value:
          type: number
        unit:
          type: string
        confidence:
          type: string
          enum: [high, medium, low]
        sourceQuote:
          type: string
        editedByUser:
          type: boolean
        originalValue:
          type: ["object", "null"]
          additionalProperties: true
      required: [id, ordinal, symbol, name, value, unit, confidence, sourceQuote, editedByUser]
    Assumption:
      type: object
      properties:
        id:
          type: integer
        ordinal:
          type: integer
        kind:
          type: string
          enum: [assumption, limitation]
        text:
          type: string
        sourceQuote:
          type: string
        confidence:
          type: string
          enum: [high, medium, low]
        editedByUser:
          type: boolean
        originalValue:
          type: ["object", "null"]
          additionalProperties: true
      required: [id, ordinal, kind, text, sourceQuote, confidence, editedByUser]
...
    ModelCard:
      type: object
      properties:
        extraction:
          type: object
          properties:
            id:
              type: integer
            projectId:
              type: integer
            sourceDocumentId:
              type: ["integer", "null"]
            providerUsed:
              type: string
              enum: [mock, openai, gemini]
            status:
              type: string
              enum: [pending, ready, failed]
            modelCardTitle:
              type: string
            domain:
              type: string
            systemDescription:
              type: string
            problemStatement:
              type: string
            odeTemplate:
              type: string
            rawExtractionJson:
              type: ["object", "null"]
              additionalProperties: true
              description: >-
                Full validated provider output matching the canonical
                ExtractionResultSchema (paper_title_or_topic, system_type,
                process_description, state_variables, parameters, equations,
                assumptions, limitations, model_card). Null for legacy rows
                created before this column existed; clients should fall back
                to the normalized tables in that case.
            providerModel:
              type: string
              description: >-
                Exact model identifier used by the provider
                (e.g. "gpt-4o", "gemini-1.5-flash", "mock").
                Empty string for legacy rows.
            systemPrompt:
              type: string
              description: >-
                System prompt sent verbatim to the AI provider.
                Contains only instructional text — NO API keys or secrets.
                Empty string for mock provider, legacy rows, and public share views.
            promptTemplateSummary:
              type: string
              description: One-line description of the extraction prompt template purpose.
            rawProviderResponse:
              type: ["object", "null"]
              additionalProperties: true
              description: >-
                Raw provider response BEFORE JSON repair and Zod validation.
                Distinct from rawExtractionJson (post-validation canonical result).
                Null for mock provider (no API call made), legacy rows, and public share views.
            repairStatus:
              type: string
              enum: [not_needed, repaired, failed]
              description: Whether the provider response required JSON repair before validation.
            validationErrors:
              type: ["string", "null"]
              description: >-
                Zod validation error details if the extraction failed or
                required repair. Null when the extraction succeeded cleanly.
            tokenUsage:
              type: ["object", "null"]
              additionalProperties: true
              description: >-
                Token count and estimated cost metadata from the provider.
                Shape varies by provider (OpenAI vs Gemini). Null when
                unavailable (mock provider always returns null).
            modelType:
              type: string
              enum:
                - chemostat
                - batch_reactor
                - fed_batch
                - cstr
                - gas_liquid_transfer
                - microalgae_pbr
                - generic_ode
              description: >-
                Auto-detected model type from the rule-based domain classifier.
                "generic_ode" is the safe default for unknown models and
                legacy rows.
            modelTypeConfidence:
              type: number
              format: float
              description: >-
                Classifier confidence score in [0, 1].
                Computed as score / (score + 10) from keyword evidence.
                0 means no domain keywords were found.
            modelTypeMatchedKeywords:
              type: array
              items:
                type: string
              description: >-
                Keywords and phrases from the source text / extracted fields
                that contributed to the auto-classification score.
            modelTypeOverride:
              type: ["string", "null"]
              enum:
                - chemostat
                - batch_reactor
                - fed_batch
                - cstr
                - gas_liquid_transfer
                - microalgae_pbr
                - generic_ode
                - null
              description: >-
                User-supplied model type override. When non-null, takes
                precedence over modelType in all frontend displays. Null means
                "use the classifier result".
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
          required:
            - id
            - projectId
            - sourceDocumentId
            - providerUsed
            - status
            - modelCardTitle
            - domain
            - systemDescription
            - problemStatement
            - odeTemplate
            - rawExtractionJson
            - providerModel
            - systemPrompt
            - promptTemplateSummary
            - rawProviderResponse
            - repairStatus
            - validationErrors
            - tokenUsage
            - modelType
            - modelTypeConfidence
            - modelTypeMatchedKeywords
            - modelTypeOverride
            - createdAt
            - updatedAt
        equations:
          type: array
          items:
            $ref: "#/components/schemas/Equation"
        variables:
          type: array
          items:
            $ref: "#/components/schemas/Variable"
        parameters:
          type: array
          items:
            $ref: "#/components/schemas/Parameter"
        assumptions:
          type: array
          items:
            $ref: "#/components/schemas/Assumption"
      required: [extraction, equations, variables, parameters, assumptions]
...